pyproject.toml
[project]
name = "zumi-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.8",
    "ipykernel>=6.29.5",
    "numpy==1.24",
    "pyserial>=3.5",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "pytest>=8.4.2",
    "httpx>=0.28.1",
    "pyexiftool==0.5",
    "av>=15.1.0",
    "tqdm>=4.67.1",
    "opencv-python==4.7.0.72",
    "scipy>=1.13.1",
    "pyyaml>=6.0.3",
    "pandas>=2.1.0",
    "scikit-fda>=0.9.1",
    "zarr>=2.18.2",
    "imagecodecs>=2024.12.30",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
